Добавил цену в продукт и IResult в сервисы

+++ OrdersApiAppPV012/Services/Repositories/CRUD/DbDaoClient.cs
@@ -5,7 +5,7 @@ using OrdersApiAppPV012.Services.Interfaces;
 
 namespace OrdersApiAppPV012.Services.Repositories.CRUD
 {
-    // имплементация dao, работающая с БД
+    // Имплементация dao, работающая с БД
     public class DbDaoClient : IDaoBase<Client>
     {
         private readonly AppDbContext db;
@@ -16,37 +16,69 @@ namespace OrdersApiAppPV012.Services.Repositories.CRUD
         }
 
 
-        // Вернуть всех клиентов
-        public async Task<IReadOnlyList<Client>> GetAllItems() =>
-            await db.Clients.ToListAsync();
+        // Вернуть список клиентов
+        public async Task<IResult> GetAllItems()
+        {
+            var clients = await db.Clients.ToListAsync();
+
+            if (clients.Count == 0)
+            {
+                return Results.NotFound(new { message = "Список пуст" });
+            }
+
+            return Results.Json(clients);
+        }
 
         // Вернуть клиента по Id
-        public async Task<Client?> GetItemById(Guid id) =>
-            await db.Clients.FirstOrDefaultAsync(c => c.Id == id);
+        public async Task<IResult> GetItemById(Guid id)
+        {
+            if (id.GetType() != typeof(Guid))
+            {
+                return Results.BadRequest(new { message = "Неверный ID" });
+            }
+
+            var client = await db.Clients.FirstOrDefaultAsync(c => c.Id == id);
+
+            if (client is null)
+            {
+                return Results.NotFound(new { message = "Нет элемента по такому ID" });
+            }
+
+            return Results.Json(client);
+        }
 
         // Добавить клиента
-        public async Task<Client> AddItem(Client client)
+        public async Task<IResult> AddItem(Client client)
         {
             await db.Clients.AddAsync(client);
             await db.SaveChangesAsync();
-            return client;
+
+            return Results.Json(client);
         }
 
         // Обновить клиента
-        public async Task<Client> UpdateItem(Client client)
+        public async Task<IResult> UpdateItem(Client client)
         {
             db.Clients.Entry(client).State = EntityState.Modified;
             await db.SaveChangesAsync();
-            return client;
+
+            return Results.Json(client);
         }
 
         // Удалить клиента
-        public async Task<bool> DeleteItem(Guid id)
+        public async Task<IResult> DeleteItem(Guid id)
         {
             var client = await db.Clients.FirstOrDefaultAsync(c => c.Id == id);
+
+            if (client is null)
+            {
+                return Results.NotFound(new { message = "Нет элемента по такому ID" });
+            }
+
             db.Clients.Remove(client);
             await db.SaveChangesAsync();
-            return true;
+
+            return Results.Ok(new { message = "Элемент удалён" });
         }
     }
 }
